Extract logic for detecting identity function
