+++ Funcky.Analyzers/Funcky.Analyzers/OptionMatchAnalyzer.cs
@@ -2,6 +2,7 @@ using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Operations;
+using static Funcky.Analyzers.IdentityFunctionMatching;
 
 namespace Funcky.Analyzers;
 
@@ -73,25 +74,4 @@ public sealed class OptionMatchAnalyzer : DiagnosticAnalyzer
 
         return null;
     }
-
-    private static bool IsIdentityFunction(IOperation operation)
-        => operation switch
-        {
-            IDelegateCreationOperation delegateCreation => IsIdentityFunction(delegateCreation.Target),
-            IAnonymousFunctionOperation anonymousFunction => IsAnonymousIdentityFunction(anonymousFunction),
-            IMethodReferenceOperation methodReference => IsFunckyIdentityFunction(methodReference),
-            _ => false,
-        };
-
-    private static bool IsAnonymousIdentityFunction(IAnonymousFunctionOperation anonymousFunction)
-        => anonymousFunction.Body.Operations.Length == 1
-           && anonymousFunction.Body.Operations[0] is IReturnOperation returnOperation
-           && anonymousFunction.Symbol.Parameters.Length == 1
-           && returnOperation.ReturnedValue is IParameterReferenceOperation;
-
-    private static bool IsFunckyIdentityFunction(IMethodReferenceOperation methodReference)
-        => methodReference.Method.Name == "Identity"
-            && SymbolEqualityComparer.Default.Equals(
-               methodReference.Method.ContainingType,
-               methodReference.SemanticModel?.Compilation.GetFunctionalType());
 }
